requirements.txt
# Master Chess workspace — analysis/monitoring tools
# The trading bot itself runs from D:\042021\CryptoBot\.venv
prometheus-client>=0.20.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
Flask>=3.0.0
